requirements.txt
streamlit
openai
python-dotenv
numpy
scikit-learn
tqdm
